pyproject.toml
[project]
name = "pdf-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "seejune2",email = "106861640+seejune2@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-openai (>=0.3.18,<0.4.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "pypdf (>=5.5.0,<6.0.0)",
    "chromadb (>=1.0.10,<2.0.0)",
    "gradio (>=5.31.0,<6.0.0)",
    "gradio-pdf (>=0.0.22,<0.0.23)",
    "langchain-text-splitters (>=0.3.8,<0.4.0)"
]

[tool.poetry]
packages = [{include = "pdf_bot", from = "src"}]


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
